requirements.txt
Flask
paddleocr
paddlepaddle
numpy
gunicorn
Pillow